♻️ refactor: 패키지 구조 변경 및 클래스 경로 수정
