+++ src/test/java/com/grow/payment_service/payment/application/service/PaymentBatchServiceImplTest.java
@@ -8,6 +8,7 @@ import java.util.List;
 
 import com.grow.payment_service.payment.application.dto.PaymentAutoChargeParam;
 import com.grow.payment_service.payment.application.dto.PaymentConfirmResponse;
+import com.grow.payment_service.payment.application.service.impl.PaymentBatchServiceImpl;
 import com.grow.payment_service.payment.domain.model.Payment;
 import com.grow.payment_service.payment.domain.model.enums.PayStatus;
 import com.grow.payment_service.payment.domain.model.PaymentHistory;
